+++ lib/simple_bpsk_SNR_qf_impl.h
@@ -27,24 +27,24 @@ namespace gr {
     class simple_bpsk_SNR_qf_impl : public simple_bpsk_SNR_qf
     {
      private:
-     unsigned int d_window_size;
+     unsigned long d_window_size;
      float d_SNR_real, d_SNR_imag, d_SNR_magn;
      
      /*! \brief computes the variance of a given number of samples with 0 mean
       */
 
-     inline void variance(float *variance, float *inputBuffer, unsigned int num_points);
+     inline void variance(float *variance, float *inputBuffer, unsigned long num_points);
      
      public:
-      simple_bpsk_SNR_qf_impl(unsigned int window_size);
+      simple_bpsk_SNR_qf_impl(unsigned long window_size);
       ~simple_bpsk_SNR_qf_impl();
 
       // getter function for the controlPorts
       float get_SNR_real() const { return d_SNR_real; }
       float get_SNR_imag() const { return d_SNR_imag; }
       float get_SNR_magn() const { return d_SNR_magn; }
-      unsigned int get_window_size() const { return d_window_size; }
-      void set_window_size(unsigned int window_size) { d_window_size = window_size; } 
+      unsigned long get_window_size() const { return d_window_size; }
+      void set_window_size(unsigned long window_size) { d_window_size = window_size; } 
       void setup_rpc();
 
       void forecast (int noutput_items, gr_vector_int &ninput_items_required);
